rename effect BlockHashRest to Rest
