--- util/isPortAvailable.ts
@@ -1,16 +1,0 @@
-export function isPortAvailable(port: number): boolean {
-  try {
-    const listener = Deno.listen({
-      transport: "tcp",
-      hostname: "127.0.0.1",
-      port,
-    });
-    listener.close();
-    return true;
-  } catch (error) {
-    if (error instanceof Deno.errors.AddrInUse) {
-      return false;
-    }
-    throw error;
-  }
-}
